backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic[email]
python-dotenv
requests
